requirements.txt
gevent==1.0.1
gevent-socketio==0.3.6
greenlet==0.4.5
pyserial==2.7
six==1.9.0
wsgiref==0.1.2
